requirements.txt
pyyaml
cdsapi
xarray
cftime
cfgrib
matplotlib
numpy
scipy
pandas
netCDF4
eccodes
pyproj
earth2studio

